Move memtile repeat and dynamic object examples from programming_examples to test/npu-xrt
