--- programming_examples/dynamic_object_fifo/lit.local.cfg
@@ -1,11 +1,0 @@
-#
-# This file is licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-#
-# (c) Copyright 2023 AMD Inc.
-
-config.suffixes = ['.lit']
-
-if 'AIE2' not in config.vitis_components:
-    config.unsupported = True
