--- programming_examples/dynamic_object_fifo/nested_loops/Makefile
@@ -1,74 +1,0 @@
-##===- Makefile -----------------------------------------------------------===##
-# 
-# This file licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-#
-# Copyright (C) 2024, Advanced Micro Devices, Inc.
-# 
-##===----------------------------------------------------------------------===##
-
-# ---
-
-# The following environment variables that point to the Xilinx runtime (XRT)
-# should be set up by an environment setup script already.
-XILINX_XRT?=/opt/xilinx/xrt
-XILINX_VITIS?=$(shell realpath $(dir $(shell which vitis))/../)
-
-# ---
-
-srcdir := $(shell dirname $(realpath $(firstword $(MAKEFILE_LIST))))
-
-XILINX_XRT_INCLUDE?=${XILINX_XRT}/include
-XILINX_XRT_LIB?=${XILINX_XRT}/lib
-
-CHESSCCWRAP2_FLAGS=aie2 -I${XILINX_VITIS}/aietools/include 
-XRT_FLAGS=-I${XILINX_XRT_INCLUDE} -L${XILINX_XRT_LIB}
-XRT_LIBS=-lxrt_coreutil
-CXX=g++-13 -ggdb 
-
-#mlir_target?=build/aie.mlir
-xclbin_target?=build/final.xclbin
-insts_target?=build/insts.txt
-host_target?=test.exe
-
-.PHONY: all
-all: ${xclbin_target} ${host_target}
-
-build/aie.mlir: ${srcdir}/aie2.py
-	mkdir -p ${@D}
-	python3 $< > $@
-
-build/kernel.o: ${srcdir}/kernel.cc
-	mkdir -p ${@D}
-	cd ${@D} && xchesscc_wrapper ${CHESSCCWRAP2_FLAGS} -c $< -o ${@F}
-
-${xclbin_target}: build/aie.mlir build/kernel.o
-	mkdir -p ${@D}
-	cd ${@D} && aiecc.py -v --aie-generate-cdo --no-compile-host --xclbin-name=${@F} \
-				--dynamic-objFifos --aie-generate-npu --npu-insts-name=${insts_target:build/%=%} ${<:%=../%}
-
-${host_target}: ${srcdir}/test.cpp ${xclbin_target}
-	rm -rf _build
-	mkdir -p _build
-	cd _build && ${powershell} cmake ${srcdir} -DTARGET_NAME=${host_target}
-	cd _build && ${powershell} cmake --build . --config Release
-ifeq "${powershell}" "powershell.exe"
-	cp _build/${host_target}.exe $@
-else
-	cp _build/${host_target} $@ 
-endif
-
-.PHONY: run
-run: ${host_target}
-	./${host_target}
-
-xclbin_sign=${XILINX_XRT}/amdxdna/setup_xclbin_firmware.sh 
-.PHONY: sign
-sign: ${xclbin_target}
-	${xclbin_sign} -dev Phoenix -xclbin $<
-
-.PHONY: clean
-clean:
-	rm -rf build _build ${targetname}
-
