--- programming_examples/dynamic_object_fifo/ping_pong/.gitignore
@@ -1,1 +1,0 @@
-build
