+++ test/npu-xrt/dynamic_object_fifo/nested_loops/kernel.cc
