+++ test/npu-xrt/dynamic_object_fifo/nested_loops/run_makefile.lit
