+++ test/npu-xrt/dynamic_object_fifo/ping_pong/kernel.cc
