+++ test/npu-xrt/dynamic_object_fifo/reduction/kernel.cc
