+++ test/npu-xrt/dynamic_object_fifo/sliding_window/kernel.cc
