+++ test/npu-xrt/dynamic_object_fifo/two_core_sliding_window/test.cpp
